+++ AI4Water/utils/datasets/datasets.py
@@ -13,7 +13,6 @@
 # https://www.bafg.de/GRDC/EN/04_spcldtbss/41_ARDB/arcticHycos.html?nn=201574 flow dataset
 # https://doi.pangaea.de/10.1594/PANGAEA.924561 Air temp Tiangin china
 
-# HYSETS https://osf.io/rpc3w/  https://www.nature.com/articles/s41597-020-00583-2
 
 # https://zenodo.org/record/3712407#.YExYDtyRWUk
 # https://zenodo.org/record/3844201#.YExYi9yRWUk
@@ -109,7 +108,6 @@
 
 # rr
 # https://zenodo.org/record/3341592#.YEx5RtyRWUk
-# https://zenodo.org/record/4539986#.YEx5T9yRWUk
 # https://zenodo.org/record/3931582#.YEx5W9yRWUk
 # https://zenodo.org/record/3528098#.YEx64NyRWUk
 
@@ -175,16 +173,14 @@ import zipfile
 import shutil, os
 from typing import Union
 
+import netCDF4
 import numpy as np
 import pandas as pd
 
-try:  # shapely may not be installed, as it may be difficult to isntall and is only needed for plotting data.
-    from AI4Water.utils.spatial_utils import plot_shapefile
-except ModuleNotFoundError:
-    plot_shapefile = None
 from AI4Water.utils.datasets.download_pangaea import PanDataSet
 from AI4Water.utils.datasets.download_zenodo import download_from_zenodo
 from AI4Water.utils.datasets.utils import download, download_all_http_directory
+from .utils import check_attributes, sanity_check
 
 
 SEP = os.sep
@@ -219,50 +215,6 @@ DATASETS = [
     'Weisssee'
             ]
 
-# following files must exist withing data folder for CAMELS-GB data
-DATA_FILES = {
-    'CAMELS-GB': [
-        'CAMELS_GB_climatic_attributes.csv',
-        'CAMELS_GB_humaninfluence_attributes.csv',
-        'CAMELS_GB_hydrogeology_attributes.csv',
-        'CAMELS_GB_hydrologic_attributes.csv',
-        'CAMELS_GB_hydrometry_attributes.csv',
-        'CAMELS_GB_landcover_attributes.csv',
-        'CAMELS_GB_soil_attributes.csv',
-        'CAMELS_GB_topographic_attributes.csv'
-    ],
-    'HYSETS': [  # following files must exist in a folder containing HYSETS dataset.
-        'HYSETS_2020_ERA5.nc',
-        'HYSETS_2020_ERA5Land.nc',
-        'HYSETS_2020_ERA5Land_SWE.nc',
-        'HYSETS_2020_Livneh.nc',
-        'HYSETS_2020_nonQC_stations.nc',
-        'HYSETS_2020_SCDNA.nc',
-        'HYSETS_2020_SNODAS_SWE.nc',
-        'HYSETS_elevation_bands_100m.csv',
-        'HYSETS_watershed_boundaries.zip',
-        'HYSETS_watershed_properties.txt'
-    ]
-}
-
-
-def gb_message():
-    link = "https://doi.org/10.5285/8344e4f3-d2ea-44f5-8afa-86d2987543a9"
-    raise ValueError(f"Dwonlaoad the data from {link} and provide the directory path as dataset=Camels(data=data)")
-
-
-def sanity_check(dataset_name, path):
-    if dataset_name in DATA_FILES:
-        if dataset_name == 'CAMELS-GB':
-            if not os.path.exists(os.path.join(path, 'data')):
-                raise FileNotFoundError(f"No folder named `data` exists inside {path}")
-            else:
-                data_path = os.path.join(path, 'data')
-                for file in DATA_FILES[dataset_name]:
-                    if not os.path.exists(os.path.join(data_path, file)):
-                        raise FileNotFoundError(f"File {file} must exist inside {data_path}")
-    return
-
 
 class Datasets(object):
 
@@ -275,6 +227,15 @@ class Datasets(object):
 
         self.units = units
         self.name = name
+    """
+    Arguments:
+        name :
+        units :
+    """
+
+    @property
+    def url(self):
+        raise NotImplementedError
 
     @property
     def base_ds_dir(self):
@@ -288,33 +249,9 @@ class Datasets(object):
             os.makedirs(_dir)
         return _dir
 
-    @property
-    def DATASETS(self):
-        raise NotImplementedError
-
-    @property
-    def location(self):  # geographical location of data
-        raise NotImplementedError
-
-    @property
-    def num_points(self):  # number of points in the data
-        raise NotImplementedError
-
-    @property
-    def start(self):  # start of data
-        raise NotImplementedError
-
-    @property
-    def end(self):  # end of data
-        raise NotImplementedError
-
-    def plot(self, points):
-        # this function should be implemented to either plot data of all points/stations or of selected points/stations.
-        raise NotImplementedError(f'The method plot should be written for the class {self.name}')
-
     def _download(self, overwrite=False):
         """Downloads the dataset. If already downloaded, then"""
-        if os.path.exists(self.ds_dir) and len(os.listdir(self.ds_dir))>0:
+        if os.path.exists(self.ds_dir) and len(os.listdir(self.ds_dir)) > 0:
             if overwrite:
                 print(f"removing previous data directory {self.ds_dir} and downloading new")
                 shutil.rmtree(self.ds_dir)
@@ -403,1415 +340,13 @@ Use overwrite=True to remove previously saved files and download again""")
         return
 
 
-class Camels(Datasets):
-
-    """
-    Get CAMELS dataset.
-    This class first downloads the CAMELS dataset if it is not already downloaded. Then the selected attribute
-    for a selected id are fetched and provided to the user using the method `fetch`.
-
-    Attributes:
-        ds_dir : diretory of the dataset
-        dynamic_attributes : tells which dynamic attributes are available in this dataset
-        static_attribute_categories : tells which kinds of static attributes are present in this category.
-
-
-    Methods:
-        stations : returns the stations for which the data (dynamic attributes) exists as list of strings.
-
-        fetch :
-            fetches all attributes (both static and dynamic type) of all station/gauge_ids
-                or a speficified station. It can also be used to fetch all attributes
-                of a number of stations ids either by providing their guage_id or
-                by just saying that we need data of 20 stations which will then
-                be chosen randomly.
-
-        fetch_dynamic_attributes :
-            fetches speficied dynamic attributes of one specified station. If the
-            dynamic attribute is not specified, all dynamic attributes will be
-            fetched for the specified station. If station is not specified, the
-            specified dynamic attributes will be fetched for all stations.
-
-        fetch_static_attributes :
-            works same as `fetch_dynamic_attributes` but for `static` attributes.
-            Here if the `category` is not specified then static attributes of
-            the specified station for all categories are returned.
-    """
-
-    DATASETS = {
-        'CAMELS-BR': {'url': "https://zenodo.org/record/3964745#.YA6rUxZS-Uk",
-                      },
-        'CAMELS-GB': {'url': gb_message},
-    }
-
-    def stations(self):
-        raise NotImplementedError
-
-    def static_attribute_categories(self):
-        raise NotImplementedError
-
-    def fetch_dynamic_attributes(self, station, dynamic_attributes, **kwargs):
-        raise NotImplementedError
-
-    def fetch_static_attributes(self, station, categories, static_attributes, as_ts=False):
-        raise NotImplementedError
-
-    @property
-    def camels_dir(self):
-        """Directory where all camels datasets will be saved. This will under datasets directory"""
-        return os.path.join(self.base_ds_dir, "CAMELS")
-
-    @property
-    def ds_dir(self):
-        """Directory where a particular dataset will be saved. """
-        return self._ds_dir
-
-    @ds_dir.setter
-    def ds_dir(self, x):
-        if x is None:
-            x = os.path.join(self.camels_dir, self.__class__.__name__)
-
-        if not os.path.exists(x):
-            os.makedirs(x)
-        #sanity_check(self.name, x)
-        self._ds_dir = x
-
-    @property
-    def common_index(self):
-        """
-        It is possible that multiple data are available during different time-periods.
-        This returns the time-pseriod/index which is shared between all the data.
-        >>>import random
-        >>>dataset = CAMELS_BR
-        >>>stations = dataset.stations()
-        >>>df = dataset.fetch_stations_attributes(station=random.choice(stations))
-        """
-        return
-
-
-    def fetch(self, stations=None,
-              dynamic_attributes='all',
-              categories='all',
-              static_attributes='all',
-              st: Union[None, str] = None,
-              en: Union[None, str] = None,
-              **kwargs)->dict:
-        """
-        Fetches the attributes of one or more stations.
-
-        Arguments:
-            stations str/list/int/float/None: default None. if string, it is supposed to be a station name/gauge_id.
-                             If list, it will be a list of station/gauge_ids. If int, it will be supposed that the
-                             user want data for this number of stations/gauge_ids. If None (default), then attributes
-                             of all available stations. If float, it will be supposed that the user wants data
-                             of this fraction of stations.
-            dynamic_attributes list: default(None), If not None, then it is the attributes to be fetched.
-                                       If None, then all available attributes are fetched
-            categories list/str: categories of static attributes to be fetched.If None, then static attributes will
-                               not be fetched.
-            static_attributes list: If categeories are available, and they are None, then this argument
-                will have no meaning.
-            st str: starting date of data to be returned. If None,
-                            the data will be returned from where it is available
-            en str: end date of data to be returned. If None, then the data will be returned till the
-                            date data is available.
-            kwargs dict: keyword arguments to read the files
-
-        returns:
-            dictionary whose keys are station/gauge_ids and values are the attributes and dataframes.
-
-        """
-        if isinstance(stations, int):
-            # the user has asked to randomly provide data for some specified number of stations
-            stations = random.sample(self.stations(), stations)
-        elif isinstance(stations, list):
-            pass
-        elif isinstance(stations, str):
-            stations = [stations]
-        elif isinstance(stations, float):
-            num_stations = int(len(self.stations()) * stations)
-            stations = random.sample(self.stations(), num_stations)
-        elif stations is None:
-            # fetch for all stations
-            stations = self.stations()
-        else:
-            raise TypeError(f"Unknown value provided for stations {stations}")
-
-        return self.fetch_stations_attributes(stations, dynamic_attributes, categories, static_attributes,
-                                              st=st, en=en,
-                                              **kwargs)
-
-    def fetch_stations_attributes(self,
-                                  stations:list,
-                                  dynamic_attributes='all',
-                                  categories=None,
-                                  static_attributes='all',
-                                  **kwargs)->dict:
-        """fetches attributes of multiple stations."""
-        assert isinstance(stations, list)
-
-        # static attributes should not be None, if we want to fetch some categories.
-        if categories:
-            assert static_attributes, f"""
-static_static attributes can not be {static_attributes}
-when categorices is {categories}"""
-
-        stations_attributes = {}
-        for station in stations:
-            stations_attributes[station] = self.fetch_station_attributes(
-                station, dynamic_attributes, categories, static_attributes, **kwargs)
-
-        return stations_attributes
-
-    def fetch_station_attributes(self,
-                                 station,
-                                 dynamic_attributes='all',
-                                 categories='all',
-                                 static_attributes='all',
-                                 as_ts=False,
-                                 st=None,
-                                 en=None,
-                                 **kwargs)->pd.DataFrame:
-        """Fetches attributes for one station.
-        Return:
-            dataframe if as_ts is True else it returns a dictionary of static and dynamic attributes for
-            a station/gauge_id
-            """
-        if st is None:
-            st = self.start
-        if en is None:
-            en = self.end
-
-        station_df = pd.DataFrame()
-        if dynamic_attributes:
-            dynamic = self.fetch_dynamic_attributes(station, dynamic_attributes, st=st, en=en, **kwargs)
-            station_df = pd.concat([station_df, dynamic])
-
-        if categories:
-            if categories == 'all':
-                categories = self.static_attribute_categories
-
-            static = self.fetch_static_attributes(station, categories, static_attributes, as_ts=as_ts)
-            if as_ts:
-                idx = pd.date_range(st, en, freq='D')
-                static = pd.DataFrame(np.repeat(static.values, len(idx), axis=0), index=idx)
-                station_df = pd.concat([station_df, static], axis=1)
-
-            else:
-                station_df = pd.concat([station_df, static])
-
-        return station_df
-
-
-class LamaH(Camels):
-    """
-    Large-Sample Data for Hydrology and Environmental Sciences for Central Europe
-    from     url = "https://zenodo.org/record/4609826#.YFNp59zt02w"
-    paper: https://essd.copernicus.org/preprints/essd-2021-72/
-    """
-    url = "https://zenodo.org/record/4609826#.YFNp59zt02w"
-    _data_types = ['total_upstrm', 'diff_upstrm_all', 'diff_upstrm_lowimp']
-
-    static_attribute_categories = ['']
-
-    def __init__(self, *, time_step:str, data_type:str, **kwargs):
-        assert time_step in ['daily', 'hourly'], f"invalid time_step {time_step} given"
-        assert data_type in self._data_types
-        self.time_step = time_step
-        self.data_type = data_type
-        super().__init__(**kwargs)
-
-        self._download()
-
-    """
-    Arguments:
-        time_step str:
-        data_type str:
-    """
-
-    @property
-    def dynamic_attributes(self):
-        station = self.stations()[0]
-        df = self.read_ts_of_station(station)
-        return df.columns.to_list()
-
-    @property
-    def static_attributes(self)->list:
-        fname = os.path.join(self.data_type_dir, f'1_attributes{SEP}Catchment_attributes.csv')
-        df = pd.read_csv(fname, sep=';', index_col='ID')
-        return df.columns.to_list()
-
-    @property
-    def ds_dir(self):
-        """Directory where a particular dataset will be saved. """
-        return os.path.join(self.camels_dir, self.name)
-
-    @property
-    def data_type_dir(self):
-        # self.ds_dir/CAMELS_AT/data_type_dir
-        f = [f for f in os.listdir(os.path.join(self.ds_dir, 'CAMELS_AT')) if self.data_type in f][0]
-        return os.path.join(self.ds_dir, f'CAMELS_AT{SEP}{f}')
-
-    def stations(self)->list:
-        # assuming file_names of the format ID_{stn_id}.csv
-        _dirs = os.listdir(os.path.join(self.data_type_dir, f'2_timeseries{SEP}{self.time_step}'))
-        s = [f.split('_')[1].split('.csv')[0] for f in _dirs]
-        return s
-
-    def fetch_station_attributes(self,
-                                 station,
-                                 dynamic_attributes='all',
-                                 categories='all',
-                                 static_attributes='all',
-                                 as_ts=False,
-                                 st=None,
-                                 en=None,
-                                 **kwargs) ->pd.DataFrame:
-        """Reads attributes of one station"""
-
-        station_df = pd.DataFrame()
-
-        if dynamic_attributes is not None:
-            dynamic_df = self.fetch_dynamic_attributes(station, dynamic_attributes, st, en, **kwargs)
-            station_df = pd.concat([station_df, dynamic_df])
-
-        if categories is not None:
-            static = self.fetch_static_attributes(station, categories, static_attributes, as_ts=as_ts)
-            if as_ts:
-                idx = pd.date_range(st, en, freq='D')
-                static = pd.DataFrame(np.repeat(static.values, len(idx), axis=0), index=idx)
-                station_df = pd.concat([station_df, static], axis=1)
-
-            else:
-                station_df = pd.concat([station_df, static])
-
-        return station_df
-
-    def fetch_static_attributes(self,
-                                station,
-                                categories,
-                                static_attributes,
-                                as_ts=False):
-
-        fname = os.path.join(self.data_type_dir, f'1_attributes{SEP}Catchment_attributes.csv')
-
-        df = pd.read_csv(fname, sep=';', index_col='ID')
-
-        static_attributes = check_attributes(static_attributes, self.static_attributes)
-
-        return df[static_attributes]
-
-    def fetch_dynamic_attributes(self,
-                                 station,
-                                 attributes='all',
-                                 st=None,
-                                 en=None,
-                                 **kwargs):
-        if st is None:
-            st = self.start
-
-        if en is None:
-            en = self.end
-
-        dynamic_attributes = check_attributes(attributes, self.dynamic_attributes)
-
-        df = self.read_ts_of_station(station)
-
-        return df[dynamic_attributes][st:en]
-
-    def read_ts_of_station(self, station)->pd.DataFrame:
-        # read a file containing timeseries data for one station
-        fname = os.path.join(self.data_type_dir, f'2_timeseries{SEP}{self.time_step}{SEP}ID_{station}.csv')
-        df = pd.read_csv(fname, sep=';')
-        periods = pd.PeriodIndex(year=df["YYYY"], month=df["MM"], day=df["DD"], freq="D")
-        df.index = periods.to_timestamp()
-        [df.pop(item) for item in ['YYYY', 'MM', 'DD']]
-        return df
-
-    @property
-    def start(self):
-        return "19810101"
-
-    @property
-    def end(self):
-        return "20191231"
-
-
-class HYSETS(Camels):
-    """
-    database for hydrometeorological modeling of 14,425 North American watersheds
-    from 1950-2018.
-    Following data_source are available.
-        SNODAS_SWE:
-        SCDNA:
-        nonQC_stations:
-        Livneh:
-        ERA5:
-    SWE data_types do not have tasmin and tasmax otherwise all datatypes have following dynamic_attributes
-    with following shapes
-        time                           (25202,)
-        watershedID                    (14425,)
-        drainage_area                  (14425,)
-        drainage_area_GSIM             (14425,)
-        flag_GSIM_boundaries           (14425,)
-        flag_artificial_boundaries     (14425,)
-        centroid_lat                   (14425,)
-        centroid_lon                   (14425,)
-        elevation                      (14425,)
-        slope                          (14425,)
-        discharge                      (14425, 25202)
-        pr                             (14425, 25202)
-        tasmax                         (14425, 25202)
-        tasmin                         (14425, 25202)
-    """
-    doi = "https://doi.org/10.1038/s41597-020-00583-2"
-    url = "https://osf.io/rpc3w/"
-    DATA_SOURCES = ['ERA5', 'ERA5Land', 'ERA5Land_SWE', 'Livneh', 'nonQC_stations', 'SCDNA', 'SNODAS_SWE']
-    def __init__(self, path, **kwargs):
-        super().__init__(**kwargs)
-        self.ds_dir = path
-
-    @property
-    def ds_dir(self):
-        return self._ds_dir
-
-    @ds_dir.setter
-    def ds_dir(self, x):
-        sanity_check('HYSETS', x)
-        self._ds_dir = x
-
-    @property
-    def static_attributes(self):
-        df = self.read_static_data()
-        return df.columns.to_list()
-
-    def stations(self)->list:
-        return self.read_static_data().index.to_list()
-
-    @property
-    def start(self):
-        return "19500101"
-
-    @property
-    def end(self):
-        return "20181231"
-
-    def fetch_static_attributes(self,
-                                station,
-                                categories='all',
-                                static_attributes='all',
-                                as_ts=False):
-
-        df = self.read_static_data()
-
-        static_attributes = check_attributes(static_attributes, self.static_attributes)
-
-        return df.loc[station][static_attributes]
-
-    def read_static_data(self):
-        fname = os.path.join(self.ds_dir, 'HYSETS_watershed_properties.txt')
-        return  pd.read_csv(fname, index_col='Watershed_ID', sep=';')
-
-
-class CAMELS_US(Camels):
-    """
-    Downloads and processes CAMELS dataset of 671 catchments named as CAMELS
-    from https://ral.ucar.edu/solutions/products/camels
-    https://doi.org/10.5194/hess-19-209-2015
-    """
-    DATASETS = ['CAMELS_US']
-    url = "https://ral.ucar.edu/sites/default/files/public/product-tool/camels-catchment-attributes-and-meteorology-for-large-sample-studies-dataset-downloads/basin_timeseries_v1p2_metForcing_obsFlow.zip"
-    catchment_attr_url = "https://ral.ucar.edu/sites/default/files/public/product-tool/camels-catchment-attributes-and-meteorology-for-large-sample-studies-dataset-downloads/camels_attributes_v2.0.zip"
-
-    folders = {'basin_mean_daymet': f'basin_mean_forcing{SEP}daymet',
-             'basin_mean_maurer': f'basin_mean_forcing{SEP}maurer',
-             'basin_mean_nldas': f'basin_mean_forcing{SEP}nldas',
-             'elev_bands': f'elev{SEP}daymet',
-             'hru': f'hru_forcing{SEP}daymet'}
-
-    def __init__(self):
-
-        super().__init__("CAMELS_US")
-
-        if os.path.exists(self.ds_dir):
-            print(f"dataset is already downloaded at {self.ds_dir}")
-        else:
-            download(self.url, os.path.join(self.camels_dir, f'CAMELS_US{SEP}CAMELS_US.zip'))
-            download(self.catchment_attr_url, os.path.join(self.camels_dir, f"CAMELS_US{SEP}catchment_attrs.zip"))
-            self._unzip()
-
-        self.attr_dir = os.path.join(self.ds_dir, f'catchment_attrs{SEP}camels_attributes_v2.0')
-        self.dataset_dir = os.path.join(self.ds_dir, f'CAMELS_US{SEP}basin_dataset_public_v1p2')
-
-    @property
-    def ds_dir(self):
-        """Directory where a particular dataset will be saved. """
-        return os.path.join(self.camels_dir, self.name)
-
-    @property
-    def start(self):
-        return "19800101"
-
-    @property
-    def end(self):
-        return "20141231"
-
-    def stations(self)->list:
-        stns = []
-        for _dir in os.listdir(os.path.join(self.dataset_dir, 'usgs_streamflow')):
-            cat = os.path.join(self.dataset_dir, f'usgs_streamflow{SEP}{_dir}')
-            stns += [fname.split('_')[0] for fname in os.listdir(cat)]
-
-        return stns
-
-    def fetch_station_attributes(self,
-                                 station,
-                                 dynamic_attributes='all',
-                                 categories='all',
-                                 static_attributes='all',
-                                 forcing='basin_mean_daymet',
-                                 as_ts=False,
-                                 **kwargs) -> pd.DataFrame:
-
-        return self.fetch_dynamic_attributes(station,
-                                             forcing=forcing,
-                                             as_ts=as_ts,
-                                             **kwargs)
-
-    def fetch_dynamic_attributes(self,
-                                 station,
-                                 dynamic_attributes='all',
-                                 forcing='basin_mean_daymet',
-                                 st=None,
-                                 en=None,
-                                 **kwargs):
-        if st is None:
-            st = self.start
-
-        if en is None:
-            en = self.end
-
-        assert isinstance(station, str)
-        df = None
-        df1 = None
-        dir_name = self.folders[forcing]
-        for cat in os.listdir(os.path.join(self.dataset_dir, dir_name)):
-            cat_dirs = os.listdir(os.path.join(self.dataset_dir, f'{dir_name}{SEP}{cat}'))
-            stn_file = f'{station}_lump_cida_forcing_leap.txt'
-            if stn_file in cat_dirs:
-                df = pd.read_csv(os.path.join(self.dataset_dir, f'{dir_name}{SEP}{cat}{SEP}{stn_file}'),
-                                 sep="\s+|;|:",
-                                 skiprows=4,
-                                 engine='python',
-                                 names=['Year', 'Mnth', 'Day', 'Hr', 'dayl(s)', 'prcp(mm/day)', 'srad(W/m2)',
-                                        'swe(mm)', 'tmax(C)', 'tmin(C)', 'vp(Pa)'],
-                                 )
-                df.index = pd.to_datetime(df['Year'].map(str) + '-' + df['Mnth'].map(str) + '-' + df['Day'].map(str))
-
-        flow_dir = os.path.join(self.dataset_dir, 'usgs_streamflow')
-        for cat in os.listdir(flow_dir):
-            cat_dirs = os.listdir(os.path.join(flow_dir, cat))
-            stn_file = f'{station}_streamflow_qc.txt'
-            if stn_file in cat_dirs:
-                fpath = os.path.join(flow_dir, f'{cat}{SEP}{stn_file}')
-                df1 = pd.read_csv(fpath,  sep="\s+|;|:'", names=['station', 'Year', 'Month', 'Day', 'Flow', 'Flag'], engine='python')
-                df1.index = pd.to_datetime(
-                    df1['Year'].map(str) + '-' + df1['Month'].map(str) + '-' + df1['Day'].map(str))
-
-        out_df = pd.concat([df[['prcp(mm/day)', 'srad(W/m2)', 'swe(mm)', 'tmax(C)', 'tmin(C)', 'vp(Pa)']], df1['Flow']],
-                           axis=1)
-        return  out_df[st:en]
-
-
-class CAMELS_BR(Camels):
-    """
-    Downloads and processes CAMELS dataset of Brazil
-    """
-    url = "https://zenodo.org/record/3964745#.YA6rUxZS-Uk"
-
-    dynamic_attributes = ['streamflow_m3s', 'streamflow_mm', 'streamflow_simulated',
-                           'precipitation_cpc', 'precipitation_mswep', 'precipitation_chirps',
-                           'evapotransp_gleam', 'evapotransp_mgb', 'potential_evapotransp_gleam',
-                           'temperature_min_cpc', 'temperature_mean', 'temperature_max']
-    def __init__(self):
-
-        super().__init__("CAMELS-BR")
-
-        self._download()
-
-    @property
-    def ds_dir(self):
-        """Directory where a particular dataset will be saved. """
-        return os.path.join(self.camels_dir, self.name)
-
-    @property
-    def static_dir(self):
-        path = None
-        for _dir in self._all_dirs:
-            if "attributes" in _dir:  # supposing that 'attributes' axist in only one file/folder in self.ds_dir
-                path = os.path.join(self.ds_dir, f'{_dir}{SEP}{_dir}')
-        return path
-
-    @property
-    def static_files(self):
-        all_files = None
-        if self.static_dir is not None:
-            all_files = glob.glob(f"{self.static_dir}/*.txt")
-        return all_files
-
-    @property
-    def static_attribute_categories(self):
-        static_attrs = []
-        for f in self.static_files:
-            ff = str(os.path.basename(f).split('.txt')[0])
-            static_attrs.append('_'.join(ff.split('_')[2:]))
-        return static_attrs
-
-    @property
-    def start(self):
-        return "19800101"
-
-    @property
-    def end(self):
-        return "20181231"
-
-    def all_stations(self, attribute)->list:
-        """Tells all station ids for which a data of a specific attribute is available."""
-        all_files = []
-        for _dir in self._all_dirs:
-            if attribute in _dir:
-                all_files = os.listdir(os.path.join(self.ds_dir, f'{_dir}{SEP}{_dir}'))
-
-        stations = []
-        for f in all_files:
-            stations.append(f.split('_')[0])
-
-        return stations
-
-    @property
-    def _all_dirs(self):
-        """All the folders in the dataset_directory"""
-        return  [f for f in os.listdir(self.ds_dir) if os.path.isdir(os.path.join(self.ds_dir, f))]
-
-    def stations(self, to_exclude=None):
-        """Returns a list of station ids which are common among all dynamic attributes.
-        >>>dataset = CAMELS_BR()
-        >>>stations = dataset.stations()
-        """
-        if to_exclude is not None:
-            if not isinstance(to_exclude, list):
-                assert isinstance(to_exclude, str)
-                to_exclude = [to_exclude]
-        else:
-            to_exclude = []
-
-        stations = {}
-        for dyn_attr in self.dynamic_attributes:
-            if dyn_attr not in to_exclude:
-                stations[dyn_attr] = self.all_stations(dyn_attr)
-
-        return list(set.intersection(*map(set, list(stations.values()) )))
-
-    def fetch_dynamic_attributes(self,
-                                 stn_id,
-                                 attributes='all',
-                                 st=None,
-                                 en=None,
-                                 **kwargs)->pd.DataFrame:
-        """
-        returns the dynamic/time series attribute/attributes for one station id.
-        >>>dataset = CAMELS_BR()
-        >>>pcp = dataset.fetch_dynamic_attributes('10500000', 'precipitation_cpc')
-        ...# fetch all time series data associated with a station.
-        >>>x = dataset.fetch_dynamic_attributes('51560000', dataset.dynamic_attributes)
-        """
-
-        if st is None:
-            st = self.start
-        if en is None:
-            en = self.end
-
-        if not kwargs:
-            kwargs = {'sep': ' '}
-
-        if attributes == 'all':
-            attributes = self.dynamic_attributes
-        elif not isinstance(attributes, list):
-            assert isinstance(attributes, str)
-            attributes = [attributes]
-
-        data = pd.DataFrame()
-        for attr in attributes:
-
-            for _dir in self._all_dirs:
-                if attr in _dir:
-                    path = os.path.join(self.ds_dir, f'{_dir}{SEP}{_dir}')
-                    # supposing that the filename starts with stn_id and has .txt extension.
-                    fname = [f for f in os.listdir(path) if f.startswith(str(stn_id)) and f.endswith('.txt')]
-                    fname = fname[0]
-                    if os.path.exists(os.path.join(path, fname)):
-                        df = pd.read_csv(os.path.join(path, fname), **kwargs)
-                        df.index = pd.to_datetime(df[['year', 'month', 'day']])
-                        df.index.freq = pd.infer_freq(df.index)
-                        df = df[st:en]
-                        [df.pop(item) for item in ['year', 'month', 'day']]
-                        data = pd.concat([data, df], axis=1)
-                    else:
-                        raise FileNotFoundError(f"file {fname} not found at {path}")
-                # else:
-                #     ValueError(f"{attributes} is not a valid dynamic attribute for {self.name}. Choose any of"
-                #            f" { self.DATASETS[self.name]['dynamic_attributes']}")
-        return data
-
-    def fetch_static_attributes(self,
-                                stn_id:int,
-                                categories='all',
-                                attributes=None,
-                                index_col_name='gauge_id',
-                                as_ts=False,
-                                **kwargs)->pd.DataFrame:
-        """
-        Arguments:
-        stn_id int:
-            station id whose attribute to fetch
-        categories str:
-            represents the type of static attribute to fetch e.g. climate, the one
-            of the file name must have this in its name, otherwise this function
-            will return None.
-        attributes str/list:
-            name of attribute to fetch. Default is None, which will return all the
-            attributes for a particular station of the specified category.
-        index_col_name str:
-            name of column containing station names
-        kwargs dict:
-            keyword arguments to be passed to pd.read_csv to read files.
-        as_ts bool:
-
-        Example
-        ```python
-        >>>dataset = Camels('CAMELS-BR')
-        >>>df = dataset.fetch_static_attributes(11500000, 'climate')
-        ```
-        """
-        if not kwargs:
-            kwargs = {'sep': ' '}
-
-        if categories == 'all':
-            categories = self.static_attribute_categories
-        elif isinstance(categories, str):
-            categories = [categories]
-
-        df = pd.DataFrame()
-
-        for category in categories:
-            fname = None
-            for f in self.static_files:
-                if category in f:
-                    fname = os.path.join(self.static_dir, f)
-                    break
-            _df = pd.read_csv(fname, **kwargs)
-
-            if attributes != 'all':
-                if isinstance(attributes, str):
-                    _attributes = [attributes]
-                else:
-                    _attributes = attributes
-            else:
-                _attributes = list(_df.columns)
-
-            _df = _df.loc[_df[index_col_name] == int(stn_id)][_attributes]
-
-            df[_attributes] = _df
-
-        return df
-
-
-class CAMELS_GB(Camels):
-    """
-    This dataset must be manually downloaded by the user.
-    The path of the downloaded folder must be provided while initiating this class.
-    """
-    def __init__(self, path=None):
-        super().__init__(name="CAMELS-GB")
-        self.ds_dir = path
-
-    @property
-    def ds_dir(self):
-        """Directory where a particular dataset will be saved. """
-        return self._ds_dir
-
-    @ds_dir.setter
-    def ds_dir(self, x):
-        sanity_check('CAMELS-GB', x)
-        self._ds_dir = x
-
-    @property
-    def static_attribute_categories(self)->list:
-        attributes = []
-        path = os.path.join(self.ds_dir, 'data')
-        for f in os.listdir(path):
-            if os.path.isfile(os.path.join(path, f)) and f.endswith('csv'):
-                attributes.append(f.split('_')[2])
-
-        return attributes
-
-    @property
-    def start(self):
-        return "19701001"
-
-    @property
-    def end(self):
-        return "20150930"
-
-    def stations(self, to_exclude=None):
-        # CAMELS_GB_hydromet_timeseries_StationID_number
-        path = os.path.join(self.ds_dir, f'data{SEP}timeseries')
-        gauge_ids = []
-        for f in os.listdir(path):
-            gauge_ids.append(f.split('_')[4])
-
-        return gauge_ids
-
-    def fetch_dynamic_attributes(self,
-                                 stn_id,
-                                 attributes='all',
-                                 st=None,
-                                 en=None,
-                                 **kwargs) ->pd.DataFrame:
-        """Fetches dynamic attribute/attributes of one station."""
-        if st is None:
-            st = self.start
-        if en is None:
-            en = self.end
-
-        path = os.path.join(self.ds_dir, f"data{SEP}timeseries")
-        fname = None
-        for f in os.listdir(path):
-            if stn_id in f:
-                fname = f
-                break
-
-        if not kwargs:
-            kwargs = {'index_col': 'date'}
-
-        df = pd.read_csv(os.path.join(path, fname), **kwargs)
-        df.index = pd.to_datetime(df.index)
-        df.index.freq = pd.infer_freq(df.index)
-        df = df[st:en]
-        if attributes != 'all':
-            return df[attributes]
-        else:
-            return df
-
-    def fetch_static_attributes(self,
-                                stn_id,
-                                categories='all',
-                                attribute='all',
-                                **kwargs)->pd.DataFrame:
-        """Fetches static attributes of one station for one or more category as dataframe."""
-
-        if categories == 'all':
-            categories = self.static_attribute_categories
-        elif isinstance(categories, str):
-            categories = [categories]
-
-        df = pd.DataFrame()
-        for category in categories:
-            cat_df = self.fetch_static_attribute(stn_id, category, **kwargs)
-            df = pd.concat([cat_df, df], axis=1)
-
-        return df
-
-    def fetch_static_attribute(self,
-                               stn_id,
-                               category,
-                               attribute='all',
-                               index_col_name='gauge_id',
-                               as_ts=False,
-                               **kwargs)->pd.DataFrame:
-
-        """Fetches static attribute/attributes of one category as dataframe."""
-        path = os.path.join(self.ds_dir, "data")
-        fname = None
-        for f in os.listdir(path):
-            if category in f:
-                fname = f
-                break
-        df = pd.read_csv(os.path.join(path, fname), **kwargs)
-        if attribute == 'all':
-            df = df.loc[df[index_col_name] == int(stn_id)]
-        else:
-            df = df.loc[df[index_col_name] == int(stn_id)][attribute]
-        if as_ts:
-            return self._to_ts(df)
-        else:
-            return df
-
-    def _to_ts(self, df):
-        return df
-
-
-class CAMELS_AUS(Camels):
-    """
-    Arguments:
-        path: path where the CAMELS-AUS dataset has been downloaded. This path must
-              contain five zip files and one xlsx file.
-    """
-    url = 'https://doi.pangaea.de/10.1594/PANGAEA.921850'
-    urls = {
-        "01_id_name_metadata.zip":"https://download.pangaea.de/dataset/921850/files/",
-        "02_location_boundary_area.zip":"https://download.pangaea.de/dataset/921850/files/",
-        "03_streamflow.zip":"https://download.pangaea.de/dataset/921850/files/",
-        "04_attributes.zip":"https://download.pangaea.de/dataset/921850/files/",
-        "05_hydrometeorology.zip":"https://download.pangaea.de/dataset/921850/files/",
-        "CAMELS_AUS_Attributes-Indices_MasterTable.csv":"https://download.pangaea.de/dataset/921850/files/",
-        "Units_01_TimeseriesData.pdf": "https://download.pangaea.de/dataset/921850/files/",
-        "Units_02_AttributeMasterTable.pdf":"https://download.pangaea.de/dataset/921850/files/",
-    }
-
-    dynamic_attributes = ['streamflow_MLd', 'streamflow_MLd_inclInfilled', 'streamflow_mmd.csv'
-                          'et_morton', 'et_morton_point_SILO',
-                          'et_morton_wet_SILO',
-                          'et_short_crop_SILO', 'et_tall_crop_SILO', 'evap_morton_lake_SILO',
-                          'evap_pan_SILO', 'evap_syn_SILO',
-                          'precipitation_AWAP', 'precipitation_SILO', 'precipitation_var_SWAP',
-                          'solarrad_AWAP', 'tmax_AWAP', 'tmin_AWAP', 'vprp_AWAP',
-                          'mslp_SILO', 'radiation_SILO', 'rh_tmax_SILO', 'rh_tmin_SILO',
-                          'tmax_SILO', 'tmin_SILO', 'vp_deficit_SILO', 'vp_SILO'
-                           ]
-
-    def __init__(self, path=None):
-        if path is not None:
-            if not os.path.exists(path) or len(os.listdir(path))<2:
-                raise FileNotFoundError(f"The path {path} does not exist")
-        self.ds_dir = path
-
-        super().__init__()
-        if not os.path.exists(self.ds_dir):
-            os.makedirs(self.ds_dir)
-
-        for _file, url in self.urls.items():
-            fpath = os.path.join(self.ds_dir, _file)
-            if not os.path.exists(fpath):
-                download(url + _file, fpath)
-
-        self._unzip()
-
-    @property
-    def start(self):
-        return "19570101"
-
-    @property
-    def end(self):
-        return "20181231"
-
-    @property
-    def location(self):
-        return "Australia"
-
-    def stations(self, as_list=True):
-        fname = os.path.join(self.ds_dir, f"01_id_name_metadata{SEP}01_id_name_metadata{SEP}id_name_metadata.csv")
-        df = pd.read_csv(fname)
-        if as_list:
-            return df['station_id'].to_list()
-        else:
-            return df
-
-    @property
-    def static_attribute_categories(self):
-        attributes = []
-        path = os.path.join(self.ds_dir, f'04_attributes{SEP}04_attributes')
-        for f in os.listdir(path):
-            if os.path.isfile(os.path.join(path, f)) and f.endswith('csv'):
-                f = str(f.split('.csv')[0])
-                attributes.append(''.join(f.split('_')[2:]))
-        return attributes
-
-    def fetch_stations_attributes(self,
-                                  stations:list,
-                                  dynamic_attributes='all',
-                                  categories=None,
-                                  static_attributes='all',
-                                  st=None,
-                                  en=None,
-                                  as_ts=False,
-                                  **kwargs) ->dict:
-        stns = {}
-        dyn_attrs = {}
-
-        if st is None:
-            st = self.start
-
-        if en is None:
-            en = self.end
-
-        if isinstance(categories, str):
-            if categories == 'all':
-                categories = self.static_attribute_categories
-            else:
-                categories = [categories]
-
-        def populate_dynattr(_path, attr):
-            fname = os.path.join(_path, attr + ".csv")
-            if os.path.exists(fname):
-                df = pd.read_csv(fname, na_values=['-99.99'], **kwargs)
-                df.index = pd.to_datetime(df[['year', 'month', 'day']])
-                [df.pop(col) for col in ['year', 'month', 'day']]
-                dyn_attrs[attr] = df
-
-        if dynamic_attributes is not None:
-            dynamic_attributes = check_attributes(dynamic_attributes, self.dynamic_attributes)
-
-            for attr in dynamic_attributes:
-                if 'SILO' in attr:
-                    path = os.path.join(self.ds_dir, f"05_hydrometeorology{SEP}05_hydrometeorology{SEP}03_Other{SEP}SILO")
-                    populate_dynattr(path, attr)
-
-                if 'AWAP' in attr:
-                    path = os.path.join(self.ds_dir, f"05_hydrometeorology{SEP}05_hydrometeorology{SEP}03_Other{SEP}AWAP")
-                    populate_dynattr(path, attr)
-
-                if 'et_' in attr or 'evap_' in attr:
-                    path = os.path.join(self.ds_dir,
-                                        f"05_hydrometeorology{SEP}05_hydrometeorology{SEP}02_EvaporativeDemand_timeseries")
-                    populate_dynattr(path, attr)
-
-                if 'precipitation_' in attr:
-                    path = os.path.join(self.ds_dir,
-                                        f"05_hydrometeorology{SEP}05_hydrometeorology{SEP}01_precipitation_timeseries")
-                    populate_dynattr(path, attr)
-
-                if 'streamflow_' in attr:
-                    path = os.path.join(self.ds_dir, f"03_streamflow{SEP}03_streamflow")
-                    populate_dynattr(path, attr)
-
-            # making one separate dataframe for one station
-            for stn in stations:
-                stn_df = pd.DataFrame()
-                for attr, attr_df in dyn_attrs.items():
-                    if attr in dynamic_attributes:
-                        stn_df[attr] = attr_df[stn][st:en]
-                stns[stn] = stn_df
-
-            if categories is not None:
-                static = self._read_static(stations, categories, static_attributes, st,en, True)
-                for k, v in stns.items():
-                    stns[k] = pd.concat([stns[k], static[k]], axis=1)
-
-        elif categories is not None:
-            return self._read_static(stations, categories, static_attributes, st, en, as_ts=as_ts)
-
-        return stns
-
-    def _read_static(self, stations, categories, attributes,
-                     st=None, en=None, as_ts=False)->dict:
-
-        df_categories = {}
-        path = os.path.join(self.ds_dir, f"04_attributes{SEP}04_attributes")
-        for category in categories:
-            for fname in os.listdir(path):
-                if category in fname:
-                    _fname = os.path.join(path, fname)
-                    df_categories[category] = pd.read_csv(_fname, index_col=['station_id'])
-
-        stn_categories = {}
-        for stn in stations:
-            stn_df = pd.DataFrame()
-            for cat, cat_df in df_categories.items():
-                stn_df = pd.concat([stn_df, cat_df.loc[cat_df.index==stn]], axis=1)
-            if attributes == 'all':
-                stn_categories[stn] = self.to_ts(stn_df, st, en, to_ts=as_ts)
-            else:
-                stn_categories[stn] = self.to_ts(stn_df[attributes], st, en, to_ts=as_ts)
-
-        return stn_categories
-
-    def to_ts(self, df:pd.DataFrame, st, en, to_ts=False):
-        if to_ts:
-            idx = pd.date_range(st, en, freq='D')
-            return pd.DataFrame(np.repeat(df.values, len(idx), axis=0), index=idx)
-        else:
-            return df
-
-    def fetch_dynamic_attributes(self,
-                                 stn_id,
-                                 attributes='all',
-                                 st=None,
-                                 en=None,
-                                 **kwargs) ->pd.DataFrame:
-        """Fetches all or selected dynamic attributes of one station."""
-
-        if st is None:
-            st = self.start
-
-        if en is None:
-            en = self.end
-
-        if attributes == 'all':
-            attributes = self.dynamic_attributes
-
-        assert isinstance(stn_id, str), f"provide only one station_id. You provided{stn_id}"
-
-        df = pd.DataFrame()
-
-        def populate_df(_path):
-            for attr in attributes:
-                fname = os.path.join(_path, attr + ".csv")
-                if os.path.exists(fname):
-                    _df = pd.read_csv(fname, usecols=[stn_id, 'year', 'month', 'day'], na_values=['-99.99'], **kwargs)
-                    _df.index = pd.to_datetime(_df[['year', 'month', 'day']].astype(str).apply(' '.join, 1))
-                    df[attr] = _df[stn_id][st:en]
-
-        if any(['SILO' in i for i in attributes]):
-            path = os.path.join(self.ds_dir, f"05_hydrometeorology{SEP}05_hydrometeorology{SEP}03_Other{SEP}SILO")
-            populate_df(path)
-
-        if any(['AWAP' in i for i in attributes]):
-            path = os.path.join(self.ds_dir, f"05_hydrometeorology{SEP}05_hydrometeorology{SEP}03_Other{SEP}AWAP")
-            populate_df(path)
-
-        if any(['et_' in i for i in attributes]) or any(['evap_' in i for i in attributes]):
-            path = os.path.join(self.ds_dir, f"05_hydrometeorology{SEP}05_hydrometeorology{SEP}02_EvaporativeDemand_timeseries")
-            populate_df(path)
-
-        if any(['precipitation_' in i for i in attributes]):
-            path = os.path.join(self.ds_dir, f"05_hydrometeorology{SEP}05_hydrometeorology{SEP}01_precipitation_timeseries")
-            populate_df(path)
-
-        if any(['streamflow_' in i for i in attributes]):
-            path = os.path.join(self.ds_dir, f"03_streamflow{SEP}03_streamflow")
-            populate_df(path)
-
-        return df
-
-    def fetch_static_attributes(self,
-                                stn_id,
-                                categories='all',
-                                attribute='all',
-                                **kwargs)->pd.DataFrame:
-        """Fetches static attribuets of one station for one or more category as dataframe."""
-
-        if categories == 'all':
-            categories = self.static_attribute_categories
-        elif isinstance(categories, str):
-            categories = [categories]
-
-        df = pd.DataFrame()
-        for category in categories:
-            cat_df = self.fetch_static_attribute(stn_id, category, **kwargs)
-            df = pd.concat([cat_df, df], axis=1)
-
-        return df
-
-    def fetch_static_attribute(self,
-                               stn_id,
-                               category,
-                               attribute='all',
-                               index_col_name='station_id',
-                               as_ts=False,
-                               **kwargs)->pd.DataFrame:
-        # fetch static attribute of one station
-        path = os.path.join(self.ds_dir, f"04_attributes{SEP}04_attributes")
-        fname = None
-        for f in os.listdir(path):
-            if category in f:
-                fname = f
-                break
-        df = pd.read_csv(os.path.join(path, fname), **kwargs)
-        if attribute == 'all':
-            df = df.loc[df[index_col_name] == stn_id]
-        else:
-            df = df.loc[df[index_col_name] == int(stn_id)][attribute]
-        if as_ts:
-            return self._to_ts(df)
-        else:
-            return df
-
-    def plot(self, what, stations=None, **kwargs):
-        assert what in ['outlets', 'boundaries']
-        f1 = os.path.join(self.ds_dir, f'02_location_boundary_area{SEP}02_location_boundary_area{SEP}shp{SEP}CAMELS_AUS_BasinOutlets_adopted.shp')
-        f2 = os.path.join(self.ds_dir, f'02_location_boundary_area{SEP}02_location_boundary_area{SEP}shp{SEP}bonus data{SEP}Australia_boundaries.shp')
-
-        if plot_shapefile is not None:
-            return plot_shapefile(f1, bbox_shp=f2, recs=stations, rec_idx=0, **kwargs)
-        else:
-            raise ModuleNotFoundError("Shapely must be installed in order to plot the datasets.")
-
-
-class CAMELS_CL(Camels):
-    """
-    Downloads and processes CAMELS dataset of Chile
-    https://doi.org/10.5194/hess-22-5817-2018
-    """
-
-    urls = {
-        "1_CAMELScl_attributes.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "2_CAMELScl_streamflow_m3s.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "3_CAMELScl_streamflow_mm.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "4_CAMELScl_precip_cr2met.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "5_CAMELScl_precip_chirps.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "6_CAMELScl_precip_mswep.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "7_CAMELScl_precip_tmpa.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "8_CAMELScl_tmin_cr2met.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "9_CAMELScl_tmax_cr2met.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "10_CAMELScl_tmean_cr2met.zip": "https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "11_CAMELScl_pet_8d_modis.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "12_CAMELScl_pet_hargreaves.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "13_CAMELScl_swe.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "14_CAMELScl_catch_hierarchy.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-        "CAMELScl_catchment_boundaries.zip":"https://store.pangaea.de/Publications/Alvarez-Garreton-etal_2018/",
-    }
-
-    dynamic_attributes = ['streamflow_m3s', 'streamflow_mm',
-                           'precip_cr2met', 'precip_chirps', 'precip_mswep', 'precip_tmpa',
-                           'tmin_cr2met', 'tmax_cr2met', 'tmean_cr2met',
-                           'pet_8d_modis', 'pet_hargreaves',
-                           'swe']
-    """
-    Arguments:
-        path: path where the CAMELS-AUS dataset has been downloaded. This path must
-              contain five zip files and one xlsx file.
-    """
-    def __init__(self,
-                 path:str = None):
-        self.ds_dir = path
-        super().__init__()
-
-        if not os.path.exists(self.ds_dir):
-            os.makedirs(self.ds_dir)
-
-        for _file, url in self.urls.items():
-            fpath = os.path.join(self.ds_dir, _file)
-            if not os.path.exists(fpath):
-                download(url+_file, fpath)
-        self._unzip()
-
-    @property
-    def _all_dirs(self):
-        """All the folders in the dataset_directory"""
-        return  [f for f in os.listdir(self.ds_dir) if os.path.isdir(os.path.join(self.ds_dir, f))]
-
-    @property
-    def start(self):
-        return "19130215"
-
-    @property
-    def end(self):
-        return "20180309"
-
-    @property
-    def static_attribute_categories(self):
-        path = os.path.join(self.ds_dir, f"1_CAMELScl_attributes{SEP}1_CAMELScl_attributes.txt")
-        df = pd.read_csv(path, sep='\t', index_col='gauge_id')
-        return df.index.to_list()
-
-    def stations(self)->list:
-        """Tells all station ids for which a data of a specific attribute is available."""
-        _stations = {}
-        for dyn_attr in self.dynamic_attributes:
-            for _dir in self._all_dirs:
-                if dyn_attr in _dir:
-                    fname = os.path.join(self.ds_dir, f"{_dir}{SEP}{_dir}.txt")
-                    df = pd.read_csv(fname, sep='\t', nrows=2, index_col='gauge_id')
-                    _stations[dyn_attr] = list(df.columns)
-
-        return list(set.intersection(*map(set, list(_stations.values()) )))
-
-    def fetch_stations_attributes(self,
-                                  stations:list,
-                                  dynamic_attributes='all',
-                                  categories=None,
-                                  static_attributes='all',
-                                  st=None,
-                                  en=None,
-                                  **kwargs) ->dict:
-        """Overwitten for speed"""
-        stns = {}
-        if st is None:
-            st = self.start
-        if en is None:
-            en = self.end
-
-        assert all(stn in self.stations() for stn in stations)
-
-        if dynamic_attributes is not None:
-            dynamic_attributes = check_attributes(dynamic_attributes, self.dynamic_attributes)
-
-            # since one file contains data for one dynamic attribute for all stations, thus if we need data
-            # for one station only, we will have to read all the files. So we first read all the files
-
-            # reading all dynnamic attributes
-            dyn_attrs = {}
-            for attr in dynamic_attributes:
-                fname = [f for f in self._all_dirs if '_'+attr in f][0]
-                fname = os.path.join(self.ds_dir, f'{fname}{SEP}{fname}.txt')
-                _df = pd.read_csv(fname, sep='\t', index_col=['gauge_id'])
-                _df.index=  pd.to_datetime(_df.index)
-                dyn_attrs[attr] = _df
-
-            # making one separate dataframe for one statino
-            for stn in stations:
-                stn_df = pd.DataFrame()
-                for attr, attr_df in dyn_attrs.items():
-                    if attr in dynamic_attributes:
-                        stn_df[attr] = attr_df[stn]
-                stns[stn] = stn_df[st:en]
-
-        if categories is not None and dynamic_attributes is None:
-            return self._read_static(stations)
-
-        return stns
-
-    def _read_static(self, stations):
-        # overwritten for speed
-        df = pd.DataFrame()
-        path = os.path.join(self.ds_dir, f"1_CAMELScl_attributes{SEP}1_CAMELScl_attributes.txt")
-        _df = pd.read_csv(path, sep='\t', index_col='gauge_id')
-
-        stns_df = {}
-        for stn in stations:
-            if stn in _df:
-                df[stn] = _df[stn]
-            elif ' ' + stn in _df:
-                df[stn] = _df[' ' + stn]
-
-            stns_df[stn] = df.transpose()
-
-        return  stns_df # to_dict('series')
-
-    def fetch_dynamic_attributes(self,
-                                 stn_id,
-                                 attributes='all',
-                                 st=None,
-                                 en=None,
-                                 **kwargs)->pd.DataFrame:
-        """Fetches dynamic attribute/attributes of one station.
-        This is provided for homogenity otherwise fetching for single station is very slow this way."""
-
-        if st is None:
-            st = self.start
-        if en is None:
-            en = self.end
-
-        if attributes == 'all':
-            attributes = self.dynamic_attributes
-
-        assert isinstance(stn_id, str), f"provide only one station_id. You provided {stn_id}"
-
-        df = pd.DataFrame(columns=attributes)
-
-        fname = None
-        for attr in attributes:
-            for f in self._all_dirs:
-                if attr in f:
-                    fname = f
-                    break
-            fpath = os.path.join(self.ds_dir, f"{fname}{SEP}{fname}.txt")
-            _df = pd.read_csv(fpath, sep='\t', index_col='gauge_id', na_values=['" "', '', ' '])
-            _df.index = pd.to_datetime(_df.index)
-            df[attr] = _df[stn_id][st:en]
-
-        return df
-
-    def fetch_static_attributes(self,
-                                station,
-                                categories=None,
-                                static_attributes=None,
-                                as_ts=False
-                                ):
-        if categories == 'all':
-            attributes = self.static_attribute_categories
-        elif categories:
-            attributes = categories
-
-        if static_attributes == 'all':
-            attributes = self.static_attribute_categories
-
-        path = os.path.join(self.ds_dir, f"1_CAMELScl_attributes{SEP}1_CAMELScl_attributes.txt")
-        try:
-            df = pd.read_csv(path, sep='\t', index_col='gauge_id', usecols=[' ' + station, 'gauge_id']).transpose()
-        except ValueError:
-            df = pd.read_csv(path, sep='\t', index_col='gauge_id', usecols=[station, 'gauge_id']).transpose()
-
-        return df[attributes]
-
-
-class HYPE(Camels):
-    """
-    Downloads and preprocesses HYPE dataset from https://zenodo.org/record/4029572.
-    This is a rainfall-runoff dataset of 564 stations from 1985 to 2019 at daily
-    monthly and yearly time steps.
-    paper : https://doi.org/10.2166/nh.2010.007
-    """
-    url = [
-        "https://zenodo.org/record/581435",
-        "https://zenodo.org/record/4029572"
-    ]
-    dynamic_attributes = [
-        'AET_mm',
-        'Baseflow_mm',
-        'Infiltration_mm',
-        'SM_mm',
-        'Streamflow_mm',
-        'Runoff_mm',
-        'Qsim_m3-s',
-        'Prec_mm',
-        'PET_mm'
-    ]
-
-    def __init__(self, time_step:str='daily', **kwargs):
-        assert time_step in ['daily', 'monthly', 'yearly']
-        self.time_step = time_step
-        self.ds_dir = None
-        super().__init__(**kwargs)
-
-        self._download()
-
-    def stations(self):
-        return np.arange(1, 565).astype(str).tolist()
-
-    def fetch_dynamic_attributes(self,
-                                 station,
-                                 attributes='all',
-                                 st=None,
-                                 en=None,
-                                 **kwargs):
-
-        dynamic_attributes = check_attributes(attributes, self.dynamic_attributes)
-
-        _dynamic_attributes = []
-        for dyn_attr in dynamic_attributes:
-            pref, suff = dyn_attr.split('_')[0], dyn_attr.split('_')[-1]
-            _dyn_attr = f"{pref}_{self.time_step}_{suff}"
-            _dynamic_attributes.append(_dyn_attr)
-
-        df = pd.DataFrame()
-        for dyn_attr in _dynamic_attributes:
-            fname = f"{dyn_attr}.csv"
-            fpath = os.path.join(self.ds_dir, fname)
-            _df = pd.read_csv(fpath, index_col='DATE', usecols=['DATE', str(station)])
-            _df.index = pd.to_datetime(_df.index)
-            _df.columns = [dyn_attr]
-            df = pd.concat([df, _df], axis=1)
-
-        return df[_dynamic_attributes][st:en]
-
-    @property
-    def start(self):
-        return '19850101'
-
-    @property
-    def end(self):
-        return '20191231'
-
-
 class Weisssee(Datasets):
 
     dynamic_attributes = ['Precipitation_measurements',
-                           'long_wave_upward_radiation',
-                           'snow_density_at_30cm',
-                           'long_wave_downward_radiation'
-                           ]
+                          'long_wave_upward_radiation',
+                          'snow_density_at_30cm',
+                          'long_wave_downward_radiation'
+                          ]
 
     url = '10.1594/PANGAEA.898217'
 
@@ -1856,8 +391,8 @@ class YamaguchiClimateJp(Weisssee):
 
 
 class FlowBenin(Weisssee):
-     """Flow data"""
-     url = "10.1594/PANGAEA.831196"
+    """Flow data"""
+    url = "10.1594/PANGAEA.831196"
 
 
 class HydrometricParana(Weisssee):
@@ -1874,7 +409,7 @@ class RiverTempSpain(Weisssee):
 class WQCantareira(Weisssee):
     """Water quality and quantity primary data from field campaigns in the Cantareira Water Supply System,
      period Oct. 2013 - May 2014"""
-    url="https://doi.pangaea.de/10.1594/PANGAEA.892384"
+    url = "https://doi.pangaea.de/10.1594/PANGAEA.892384"
 
 
 class RiverIsotope(Weisssee):
@@ -1886,6 +421,7 @@ class EtpPcpSamoylov(Weisssee):
     """Evpotranspiration and Precipitation at station TOWER on Samoylov Island Russia
      from 20110524 to 20110819 with 30 minute frequency"""
     url = "10.1594/PANGAEA.811076"
+
 
 class FlowSamoylov(Weisssee):
     """Net lateral flow at station INT2 on Samoylov Island Russia
@@ -1989,10 +525,10 @@ class WeatherJena(Datasets):
         unzip_all_in_dir(sub_dir, 'zip')
 
     def fetch(self,
-              st:str=None,
-              en:str=None,
+              st: str = None,
+              en: str = None,
               **kwargs
-              )->pd.DataFrame:
+              ) -> pd.DataFrame:
 
         sub_dir = os.path.join(self.ds_dir, self.obs_loc)
         all_files = glob.glob(f"{sub_dir}/*.csv")
@@ -2000,7 +536,7 @@ class WeatherJena(Datasets):
         df = pd.DataFrame()
         for fpath in all_files:
             f_df = pd.read_csv(fpath, index_col='Date Time',
-                               encoding= 'unicode_escape', na_values=-9999)
+                               encoding = 'unicode_escape', na_values=-9999)
             f_df.index = pd.DatetimeIndex(f_df.index)
             df = pd.concat([df, f_df])  # todo, such concatenation is slow.
 
@@ -2014,10 +550,164 @@ class WeatherJena(Datasets):
         return df[st:en]
 
 
+class SWECanada(Datasets):
+    """
+    Daily Canadian historical Snow Water Equivalent dataset from 1928 to 2020
+    https://doi.org/10.1080/07055900.2019.1598843
+    """
+    url = "https://doi.org/10.5194/essd-2021-160"
+    feaures = ['snw', 'snd', 'den']
+    q_flags = ['data_flag_snw', 'data_flag_snd', 'qc_flag_snw', 'qc_flag_snd']
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+        self._download()
+
+    def stations(self) -> list:
+        nc = netCDF4.Dataset(os.path.join(self.ds_dir, 'CanSWE-CanEEN_1928-2020_v1.nc'))
+        s = nc['station_id'][:]
+        return s.tolist()
+
+    @property
+    def start(self):
+        return '19280101'
+
+    @property
+    def end(self):
+        return '20200731'
+
+    def fetch(self,
+              station_id: Union[None, str, float, int, list] = None,
+              features: Union[None, str, list] = None,
+              q_flags: Union[None, str, list] = None,
+              st=None,
+              en=None,
+              ) -> dict:
+        """
+        Fetches time series data from selected stations.
+        Arguments:
+            station_id : station/stations to be retrieved. In None, then data
+                          from all stations will be returned.
+            features : Names of features to be retrieved. Following features
+                are allowed:
+                    - 'snw' snow water equivalent kg/m3
+                    - 'snd' snow depth m
+                    - 'den' snowpack bulk density kg/m3
+                If None, then all three features will be retrieved.
+            q_flags : If None, then no qflags will be returned. Following q_flag
+                values are available.
+                    - 'data_flag_snw'
+                    - 'data_flag_snd'
+                    - 'qc_flag_snw'
+                    - 'qc_flag_snd'
+            st : start of data to be retrieved
+            en : end of data to be retrived.
+        returns:
+            a dictionary of dataframes of shape (st:en, features + q_flags) whose
+            length is equal to length of stations being considered.
+        """
+
+        if station_id is None:
+            station_id - self.stations()
+        elif isinstance(station_id, str):
+            station_id = [station_id]
+        elif isinstance(station_id, list):
+            pass
+        elif isinstance(station_id, int):
+            station_id = random.sample(self.stations(), station_id)
+        elif isinstance(station_id, float):
+            num_stations = int(len(self.stations()) * station_id)
+            station_id = random.sample(self.stations(), num_stations)
+
+        stns = self.stations()
+        stn_id_dict = {k: v for k, v in zip(stns, np.arange(len(stns)))}
+        stn_id_dict_inv = {v: k for k, v in stn_id_dict.items()}
+        stn_ids = [stn_id_dict[i] for i in station_id]
+
+        features = check_attributes(features, self.feaures)
+        qflags = []
+        if q_flags is not None:
+            qflags = check_attributes(q_flags, self.q_flags)
+
+        features_to_fetch = features + qflags
+
+        all_stn_data = {}
+        for stn in stn_ids:
+
+            stn_df = self.fetch_station_attributes(stn, features_to_fetch, st=st, en=en)
+            all_stn_data[stn_id_dict_inv[stn]] = stn_df
+
+        return all_stn_data
+
+    def fetch_station_attributes(self,
+                                 stn,
+                                 features_to_fetch,
+                                 st=None,
+                                 en=None,
+                                 ) -> pd.DataFrame:
+        """fetches attributes of one station"""
+
+        st, en = self._check_length(st, en)
+
+        nc = netCDF4.Dataset(os.path.join(self.ds_dir, 'CanSWE-CanEEN_1928-2020_v1.nc'))
+
+        stn_df = pd.DataFrame(columns=features_to_fetch)
+
+        for var in nc.variables:
+            if var in features_to_fetch:
+                ma = np.array(nc[var][:])
+                ma[ma == nc[var]._FillValue] = np.nan
+                ta = ma[stn, :]  # target array of on station
+                s = pd.Series(ta, index=pd.date_range(self.start, self.end, freq='D'), name=var)
+                stn_df[var] = s[st:en]
+
+        nc.close()
+
+        return stn_df
+
+
+class RRLuleaSweden(Datasets):
+    """
+    Rainfall runoff data for an urban catchment from 2016-2019
+    https://doi.org/10.5194/hess-24-869-2020
+    """
+    url = "https://zenodo.org/record/3931582"
+
+
+class RRAlpileCatchments(Datasets):
+    """
+    Modelled runoff in contrasting Alpine catchments in Austria from 1981 to 2100
+    using 14 models.
+    past 1981 - 2010
+    future
+    https://hess.copernicus.org/preprints/hess-2021-92/
+    """
+    url = "https://zenodo.org/record/4539986"
+
+
+class ETPAgroForestGermany(Datasets):
+    """
+    Evapotranspiration over agroforestry sites in Germany
+    https://doi.org/10.5194/bg-17-5183-2020
+    SiteName_Landuse_Content_Figures_Tables.csv
+    """
+    url = "https://zenodo.org/record/4038399"
+
+
+class ETPTelesinaItaly(Datasets):
+    """
+    Daily rain and reference evapotranspiration for three years 2002-2004
+    """
+    url = "https://zenodo.org/record/3726856"
+
+
 class Laos(Datasets):
     """
     Downloads and prepares hydrological, climate and land use data for Laos.
     """
+    inputs = ['prec', 'et']
+    target = ['e_coli']
 
     def fetch_lu(self):
         """Downloads and unzips and landuse data"""
@@ -2026,20 +716,33 @@ class Laos(Datasets):
         download(url, fname)
         shutil.unpack_archive(fname, self.ds_dir)
 
+    def fetch(self,
+              inputs: Union[None, list],
+              target: Union[str, list],
+              hru_definition: Union[None, str],
+              st: Union[None, str],
+              en: Union[None, str],
+              **kwargs):
+        """
+        Fetches dataset of Laos for rainfall runoff modeling [1].
+        The user can specify `hru_defintion` wose details can be seen in [2].
 
-def check_attributes(attributes, check_against:list)->list:
-    if attributes == 'all':
-        attributes = check_against
-    elif not isinstance(attributes, list):
-        assert isinstance(attributes, str)
-        assert attributes in check_against
-        attributes = [attributes]
-    else:
-        assert isinstance(attributes, list), f'unknown attributes {attributes}'
+        Arguments:
+            inputs list: names of inputs to use.
+            target str/list:
+            hru_definition str:
+            st str:
+            en str:
+            **kwargs dict:
 
-    assert all(elem in check_against for elem in attributes)
+        returns:
+            a dataframe of shape (inputs+target, st:en)
+        """
+        inputs = check_attributes(inputs, self.inputs)
+        target = check_attributes(target, self.target)
+        features_to_fecth = inputs + target
 
-    return attributes
+        return
 
 
 def unzip_all_in_dir(dir_name, ext=".gz"):
